[refactor] 매크로 대시보드 UI 책임 분리 #366
